remove micro data lib
