+++ packages/ndla-licenses/src/CCRel.js
@@ -31,41 +31,10 @@ export const getResourceTypeNamespace = type => {
   }
 };
 
-export const microDataTypes = {
-  person: 'person',
-  organization: 'organization',
-  thing: 'thing',
-  url: 'url',
-  creativeWork: 'creativeWork',
-};
-
-export const getMicroDataNamespaceByType = type => {
-  if (!type) throw new Error('type is not defined');
-  const capitalized = type.charAt(0).toUpperCase() + type.slice(1);
-  return `http://schema.org/${capitalized}`;
-};
-
 export const metaTypes = {
   author: 'author',
   copyrightHolder: 'copyrightHolder',
   contributor: 'contributor',
   title: 'title',
   other: 'other',
-};
-
-export const getMicroDataNamespaceByMicroDataTypeWithFallback = (
-  metaType,
-  microDataType = null,
-) => {
-  if (microDataType) {
-    return getMicroDataNamespaceByType(microDataType);
-  }
-
-  let fallbackMicroDataType = microDataTypes.person;
-
-  if (metaType === metaTypes.copyrightHolder) {
-    fallbackMicroDataType = microDataTypes.organization;
-  }
-
-  return getMicroDataNamespaceByType(fallbackMicroDataType);
 };
