+++ packages/ndla-licenses/src/__tests__/CCRel-test.js
@@ -1,0 +1,24 @@
+/**
+ * Copyright (c) 2016-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+/* eslint-env jest */
+
+import {
+  resourceTypes,
+  getResourceTypeNamespace,
+} from '../CCRel';
+
+test('microData/getResourceTypeNamespace when type video should return video namespace', () => {
+  const namespace = getResourceTypeNamespace(resourceTypes.video);
+  expect(namespace).toBe('http://purl.org/dc/dcmitype/MovingImage');
+});
+
+test('microData/getResourceTypeNamespace when unknown type should return null', () => {
+  const namespace = getResourceTypeNamespace('UNKNOWN');
+  expect(namespace).toBe(null);
+});
